Moved ShieldedContext into shared crate.
